Move bootstrap.py into streamlit.web
